added option to read from ts config file
