+++ cli/defineConfig.js
@@ -1,0 +1,15 @@
+"use strict";
+// Note: This file is also being used by the dev cli
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.defineConfig = void 0;
+/**
+ * Define a configuration file to run We in development modes with groups
+ * and applets pre-installed.
+ *
+ * @param config
+ * @returns
+ */
+function defineConfig(config) {
+    return config;
+}
+exports.defineConfig = defineConfig;
