+++ package.json
@@ -30,7 +30,7 @@
     "build:applet-iframe": "yarn workspace applet-iframe build",
     "build:happ-iframe": "yarn workspace happ-iframe build",
     "build:utils": "yarn workspace applet-iframe install && yarn build:applet-iframe && yarn build:happ-iframe",
-    "build:cli": "rimraf ./cli/dist && cp -r ./out ./cli && mv ./cli/out ./cli/dist && cp ./cli/cli.js ./cli/dist && cp moss.config.json ./cli/dist/main && cp -r ./resources ./cli/dist/main/resources && rimraf ./cli/dist/main/resources/bins",
+    "build:cli": "rimraf ./cli/dist && cp -r ./out ./cli && mv ./cli/out ./cli/dist && cp moss.config.json ./cli/dist/main && cp -r ./resources ./cli/dist/main/resources && rimraf ./cli/dist/main/resources/bins && tsc -p tsconfig.defineConfig.json",
     "publish:applets": "run-singleton \"node scripts/publish-applets.mjs\"",
     "prepare:default-apps": "yarn prepare:happ && yarn fetch:default-apps",
     "fetch:default-apps": "yarn download:appstore && yarn download:kando",
@@ -61,12 +61,14 @@
     "mime": "4.0.0",
     "nanoid": "5.0.4",
     "split": "1.0.1",
+    "ts-node": "10.9.2",
     "winston": "3.11.0"
   },
   "devDependencies": {
     "@electron-toolkit/eslint-config-prettier": "^1.0.1",
     "@electron-toolkit/eslint-config-ts": "^1.0.0",
     "@electron-toolkit/tsconfig": "^1.0.1",
+    "@lightningrodlabs/we-dev-cli": "file:./cli",
     "@types/commander": "2.12.2",
     "@types/nanoid": "3.0.0",
     "@types/node": "^18.17.5",
