+++ tsconfig.node.json
@@ -1,6 +1,6 @@
 {
   "extends": "./node_modules/@electron-toolkit/tsconfig/tsconfig.node.json",
-  "include": ["electron.vite.config.*", "src/main/*", "src/preload/*"],
+  "include": ["electron.vite.config.*", "src/main/*", "src/preload/*", "src/main/cli/*"],
   "compilerOptions": {
     "composite": true
   }
